Move Posto dataclass to dataclasses module
